requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.9.3
django-cors-headers==2.4.0
djangorestframework==3.6.4
httpie==0.9.9
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
requests==2.19.1
urllib3==1.23
